Statistics Query added with factory pattern
