+++ res/static/js/lib/underscore.js
@@ -139,7 +139,7 @@
     return typeof length == 'number' && length >= 0 && length <= MAX_ARRAY_INDEX;
   };
 
-  // Collection Functions
+  // collection Functions
   // --------------------
 
   // The cornerstone, an `each` implementation, aka `forEach`.
