+++ src/jonamatoka/violet/util/qstat/collection/ProductsById.java
@@ -1,8 +1,7 @@
-package jonamatoka.violet.util.Statisticsnewversion.Collection;
+package jonamatoka.violet.util.qstat.collection;
 
 import jonamatoka.violet.data.model.Product;
 import jonamatoka.violet.data.repo.ProductRepository;
-import jonamatoka.violet.util.Statisticsnewversion.Collection.Collection;
 
 import java.util.ArrayList;
 import java.util.List;
