+++ src/jonamatoka/violet/util/qstat/collection/ProductsByOwnerId.java
@@ -1,24 +1,23 @@
-package jonamatoka.violet.util.Statisticsnewversion.Collection;
+package jonamatoka.violet.util.qstat.collection;
 
 import jonamatoka.violet.data.model.Product;
 import jonamatoka.violet.data.model.ProductStack;
 import jonamatoka.violet.data.model.Store;
 import jonamatoka.violet.data.repo.ProductRepository;
 import jonamatoka.violet.data.repo.StoreRepository;
-import jonamatoka.violet.util.Statisticsnewversion.Collection.Collection;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class ProductsByStoreId implements Collection<Product> {
+public class ProductsByOwnerId implements Collection<Product> {
 
-    private long storeId;
+    private String ownerId;
     private StoreRepository storeRepository=null;
     private ProductRepository productRepository=null;
 
-    public ProductsByStoreId(StoreRepository storeRepository, ProductRepository productRepository, long storeId){
+    public ProductsByOwnerId(StoreRepository storeRepository, ProductRepository productRepository, String ownerId){
 
-        this.storeId = storeId;
+        this.ownerId=ownerId;
         this.storeRepository=storeRepository;
         this.productRepository=productRepository;
 
@@ -28,13 +27,14 @@ public class ProductsByStoreId implements Collection<Product> {
     public List<Product> calc() {
 
         List products = new ArrayList<Product>();
-        Store s=storeRepository.findOne(storeId);
-        for(ProductStack ps : s.getInventory().get()){
-            products.add(productRepository.findOne(ps.getProductId()));
+        List <Store> storesOwned = new ArrayList<>(new StoresByOwnerId(storeRepository,ownerId).calc());
+        for(Store s:storesOwned) {
+            for (ProductStack ps : s.getInventory().get()) {
+                products.add(productRepository.findOne(ps.getProductId()));
+            }
         }
-
         return products;
-
     }
 
 }
+
